refactor: 공지컴포넌트 모듈화 resolve #29
