+++ src/components/index.js
@@ -14,3 +14,5 @@ export { default as Teacher } from './layouts/teacher';
 export { default as Director } from './layouts/director';
 
 export { default as TransferList } from './transferList';
+
+export { default as Notice } from './notice/notice';
